+++ examples/vue-app/src/chains/solana.vue
@@ -32,7 +32,7 @@ import {
 import { OpenloginAdapter } from "@web3auth/openlogin-adapter";
 import { InjectedProvider, TorusInjectedProvider } from "@web3auth/solana-provider";
 import { SolanaWalletConnectorPlugin } from "@web3auth/solana-wallet-connector-plugin";
-import { SolletAdapter, SolletExtensionAdapter } from "@web3auth/sollet-adapter";
+import { SolletWebAdapter, SolletExtensionAdapter } from "@web3auth/sollet-adapter";
 import { Web3Auth } from "@web3auth/web3auth";
 import Vue from "vue";
 
@@ -127,7 +127,7 @@ export default Vue.extend({
             clientId: config.clientId,
           },
         });
-        const solletAdapter = new SolletAdapter();
+        const solletAdapter = new SolletWebAdapter();
         const solletExtensionAdapter = new SolletExtensionAdapter();
         this.web3auth.configureAdapter(solletAdapter);
         this.web3auth.configureAdapter(solletExtensionAdapter);
